refactor: move LaunchAppButton to common components folder
